Refactor backend package (#56)
* Refactor backend package

* Update Makefile
