+++ gnark_backend_ffi/backend/groth16/raw_gate.go
@@ -1,4 +1,4 @@
-package structs
+package groth16
 
 import (
 	"encoding/json"
@@ -8,17 +8,17 @@ import (
 	fr_bn254 "github.com/consensys/gnark-crypto/ecc/bn254/fr"
 )
 
-type ArithmeticOpcode struct {
-	MulTerms []backend.MulTerm
-	AddTerms []backend.AddTerm
-	qM       fr_bn254.Element
+type RawGate struct {
+	MulTerms     []backend.MulTerm
+	AddTerms     []backend.AddTerm
+	ConstantTerm fr_bn254.Element
 }
 
-func (g *ArithmeticOpcode) UnmarshalJSON(data []byte) error {
-	var gateMap map[string]interface{}
-	err := json.Unmarshal(data, &gateMap)
+func (g *RawGate) UnmarshalJSON(data []byte) error {
+	var rawGateMap map[string]interface{}
+	err := json.Unmarshal(data, &rawGateMap)
 	if err != nil {
-		log.Print(err)
+		log.Fatal(err)
 		return err
 	}
 
@@ -27,50 +27,50 @@ func (g *ArithmeticOpcode) UnmarshalJSON(data []byte) error {
 	var constantTerm fr_bn254.Element
 
 	// Deserialize mul terms.
-	if mulTermsValue, ok := gateMap["mul_terms"].([]interface{}); ok {
+	if mulTermsValue, ok := rawGateMap["mul_terms"].([]interface{}); ok {
 		mulTermsJSON, err := json.Marshal(mulTermsValue)
 		if err != nil {
-			log.Print(err)
+			log.Fatal(err)
 			return err
 		}
 		err = json.Unmarshal(mulTermsJSON, &mulTerms)
 		if err != nil {
-			log.Print(err)
+			log.Fatal(err)
 			return err
 		}
 	} else {
-		log.Print("Error: couldn't deserialize mul terms.")
+		log.Fatal("Error: couldn't deserialize mul terms.")
 		return &json.UnmarshalTypeError{}
 	}
 
 	// Deserialize add terms.
-	if addTermsValue, ok := gateMap["add_terms"].([]interface{}); ok {
+	if addTermsValue, ok := rawGateMap["add_terms"].([]interface{}); ok {
 		addTermsJSON, err := json.Marshal(addTermsValue)
 		if err != nil {
-			log.Print(err)
+			log.Fatal(err)
 			return err
 		}
 		err = json.Unmarshal(addTermsJSON, &addTerms)
 		if err != nil {
-			log.Print(err)
+			log.Fatal(err)
 			return err
 		}
 	} else {
-		log.Print("Error: couldn't deserialize add terms.")
+		log.Fatal("Error: couldn't deserialize add terms.")
 		return &json.UnmarshalTypeError{}
 	}
 
 	// Deserialize constant term.
-	if encodedConstantTerm, ok := gateMap["constant_term"].(string); ok {
+	if encodedConstantTerm, ok := rawGateMap["constant_term"].(string); ok {
 		constantTerm = backend.DeserializeFelt(encodedConstantTerm)
 	} else {
-		log.Print("Error: coefficient is not a felt.")
+		log.Fatal("Error: coefficient is not a felt.")
 		return &json.UnmarshalTypeError{}
 	}
 
 	g.MulTerms = mulTerms
 	g.AddTerms = addTerms
-	g.qM = constantTerm
+	g.ConstantTerm = constantTerm
 
 	return nil
 }
