+++ gnark_backend_ffi/backend/groth16/raw_gate_test.go
@@ -1,4 +1,4 @@
-package structs
+package groth16
 
 import (
 	"encoding/json"
