+++ gnark_backend_ffi/backend/plonk/black_box_function_call_opcode.go
@@ -1,4 +1,4 @@
-package structs
+package groth16
 
 import "gnark_backend_ffi/backend"
