+++ gnark_backend_ffi/backend/plonk/directive_opcode.go
@@ -1,4 +1,4 @@
-package structs
+package groth16
 
 import (
 	"encoding/json"
