+++ gnark_backend_ffi/backend/plonk/directive_opcode_test.go
@@ -1,4 +1,4 @@
-package structs
+package groth16
 
 import (
 	"encoding/json"
