+++ gnark_backend_ffi/backend/plonk/opcode.go
@@ -1,3 +1,3 @@
-package structs
+package groth16
 
 type Opcode interface{}
